build: move feature flag scripts from package to root
