+++ .github/workflows/publish_experimental_release.yml
@@ -17,7 +17,6 @@ jobs:
           cache: 'yarn'
       - run: |
           yarn install
-          yarn workspace @pluralsight/headless-styles up @pluralsight/shared@experimental
           yarn build:packages
       - name: Archive production artifacts
         uses: actions/upload-artifact@v3
