+++ .pnp.cjs
@@ -37,10 +37,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:packages/react-utils"\
       },\
       {\
-        "name": "@pluralsight/shared",\
-        "reference": "workspace:packages/shared"\
-      },\
-      {\
         "name": "release-script",\
         "reference": "workspace:scripts"\
       },\
@@ -60,7 +56,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@pluralsight/headless-styles", ["workspace:packages/headless-styles"]],\
       ["@pluralsight/icons", ["workspace:packages/icons"]],\
       ["@pluralsight/react-utils", ["workspace:packages/react-utils"]],\
-      ["@pluralsight/shared", ["workspace:packages/shared"]],\
       ["@pluralsight/tva", ["workspace:."]],\
       ["release-script", ["workspace:scripts"]],\
       ["sandbox", ["workspace:packages/headless-styles/sandbox"]],\
@@ -988,10 +983,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       [\
         "@pluralsight/react-utils",\
         "virtual:8693780468136bf004f9b532d653844c342247b6a39276486a4e8cbecc2c295c87d04c42b26f9c469bc9dfcf5a0d1f6012e10b6c148cd44051b30c652ede2384#npm:0.0.0-experimental-82737f-20220825"\
-      ],\
-      [\
-        "@pluralsight/shared",\
-        "workspace:packages/shared"\
       ],\
       [\
         "@polka/url",\
@@ -18809,7 +18800,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageDependencies": [\
             ["@pluralsight/headless-styles", "workspace:packages/headless-styles"],\
             ["@pluralsight/design-tokens", "npm:0.0.0-experimental-82737f-20220825"],\
-            ["@pluralsight/shared", "workspace:packages/shared"],\
             ["@types/copyfiles", "npm:2.4.1"],\
             ["@types/jest", "npm:28.1.6"],\
             ["@types/node", "npm:18.7.14"],\
@@ -18906,17 +18896,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["renamer", "npm:4.0.0"],\
             ["tslib", "npm:2.4.0"],\
             ["typescript", "patch:typescript@npm%3A4.7.4#~builtin<compat/typescript>::version=4.7.4&hash=7ad353"]\
-          ],\
-          "linkType": "SOFT"\
-        }]\
-      ]],\
-      ["@pluralsight/shared", [\
-        ["workspace:packages/shared", {\
-          "packageLocation": "./packages/shared/",\
-          "packageDependencies": [\
-            ["@pluralsight/shared", "workspace:packages/shared"],\
-            ["chalk", "npm:5.0.1"],\
-            ["replace-in-file", "npm:6.3.5"]\
           ],\
           "linkType": "SOFT"\
         }]\
@@ -22973,21 +22952,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ]],\
             ["ansi-styles", "npm:4.3.0"],\
             ["supports-color", "npm:7.2.0"]\
-          ],\
-          "linkType": "HARD"\
-        }],\
-        ["npm:5.0.1", {\
-          "packageLocation": "./.yarn/cache/chalk-npm-5.0.1-6afcb94227-7b45300372.zip/node_modules/chalk/",\
-          "packageDependencies": [\
-            ["chalk", "npm:5.0.1"],\
-            ["#ansi-styles", [\
-              "ansi-styles",\
-              "npm:6.1.0"\
-            ]],\
-            ["#supports-color", [\
-              "supports-color",\
-              "npm:9.2.2"\
-            ]]\
           ],\
           "linkType": "HARD"\
         }]\
