+++ packages/headless-styles/package.json
@@ -46,7 +46,6 @@
     "test": "echo \"'yarn test' should be run from root directory.\" && exit 1"
   },
   "dependencies": {
-    "@pluralsight/shared": "^0.0.0",
     "kebab-case": "^1.0.1",
     "tslib": "latest"
   },
