+++ packages/headless-styles/src/globalJS.d.ts
@@ -1,2 +1,1 @@
 declare module 'kebab-case'
-declare module '@pluralsight/shared'
