--- packages/shared/scripts/theme.mjs
@@ -1,13 +1,0 @@
-#!/usr/bin/env node
-
-'use strict'
-
-import chalk from 'chalk'
-
-export const command = chalk.whiteBright
-export const error = chalk.bold.red
-export const info = chalk.bold.cyan
-export const importantNote = chalk.bold.yellow
-export const link = chalk.underline.blueBright
-export const success = chalk.bold.greenBright
-export const warning = chalk.redBright
