--- packages/shared/src/index.js
@@ -1,1 +1,0 @@
-export * from './featureFlags'
